backend/requirements.txt
# FastAPI & Server
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-multipart>=0.0.9

# Database
sqlalchemy[asyncio]>=2.0.0
asyncpg>=0.29.0
alembic>=1.13.0
psycopg2-binary>=2.9.9

# Pydantic
pydantic[email]>=2.0.0
pydantic-settings>=2.0.0
email-validator>=2.0.0

# Google Cloud & AI
google-cloud-storage>=2.18.0
google-genai>=1.0.0

# Authentication
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

# Utilities
python-dotenv>=1.0.0
httpx>=0.27.0

# PDF Processing
pypdf>=4.0.0
python-docx>=1.1.0

# Development
pytest>=8.0.0
pytest-asyncio>=0.23.0
